+++ common/requests/mutate.ts
@@ -2,14 +2,13 @@
  * Functions to mutate/change data on the API
  */
 import { AuthSession, CreatedPost } from "../types";
-import { getAuthSession } from "../utils";
 import doRequest from "./request";
 
-export async function login(email: string, password: string) {
+export async function login(userId: string, email: string, password: string) {
   return await doRequest<AuthSession>(
     "/api/user/login",
     {
-      userId: (await getAuthSession()).userId,
+      userId,
       email,
       password,
     },
@@ -21,7 +20,6 @@ export async function register(name: string, email: string, password: string) {
   await doRequest(
     "/api/user/register",
     {
-      userId: (await getAuthSession()).userId,
       name,
       email,
       password,
@@ -30,11 +28,11 @@ export async function register(name: string, email: string, password: string) {
   );
 }
 
-export async function createPost(prompt: string) {
+export async function createPost(userId: string, prompt: string) {
   return await doRequest<CreatedPost>(
     "/api/post/create",
     {
-      userId: (await getAuthSession()).userId,
+      userId: userId,
       prompt,
     },
     "POST"
