--- common/utils/isReactNative.ts
@@ -1,3 +1,0 @@
-export function isReactNative() {
-  return typeof document === "undefined";
-}
