+++ mobile/package.json
@@ -15,33 +15,33 @@
     ]
   },
   "dependencies": {
-    "@daydream/common": "1.0.0",
-    "@expo/webpack-config": "^0.17.0",
+    "@daydream/common": "workspace:^",
+    "@expo/webpack-config": "0.17.3",
     "@react-navigation/native": "^6.0.13",
     "@react-navigation/stack": "^6.3.3",
-    "@types/react": "~18.0.0",
+    "@types/react": "18.0.25",
     "@types/react-native": "~0.69.1",
-    "expo": "~46.0.13",
+    "expo": "47.0.3",
     "expo-secure-store": "^11.3.0",
     "expo-status-bar": "~1.4.0",
     "jest": "^26.6.3",
     "jest-expo": "^46.0.1",
     "jsonwebtoken": "^8.5.1",
     "nativewind": "^2.0.11",
-    "react": "18.0.0",
-    "react-dom": "18.0.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-native": "0.69.6",
     "react-native-gesture-handler": "^2.8.0",
     "react-native-safe-area-context": "4.3.1",
     "react-native-screens": "~3.15.0",
     "react-native-vector-icons": "^9.2.0",
     "react-native-web": "~0.18.7",
-    "typescript": "^4.6.3"
+    "typescript": "4.8.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@types/react-native-vector-icons": "^6.4.12",
-    "tailwindcss": "^3.2.1"
+    "@babel/core": "7.20.2",
+    "@types/react-native-vector-icons": "6.4.12",
+    "tailwindcss": "3.2.4"
   },
   "private": true,
   "packageManager": "yarn@3.2.4"
