+++ web/package.json
@@ -10,17 +10,19 @@
     "test": "bash ./run_tests.sh"
   },
   "dependencies": {
-    "@daydream/common": "1.0.0",
-    "@headlessui/react": "1.7.3",
-    "@prisma/client": "4.6.0",
-    "@tanstack/react-query": "4.12.0",
+    "@daydream/common": "workspace:^",
+    "@headlessui/react": "1.7.4",
+    "@prisma/client": "4.6.1",
+    "@tanstack/react-query": "4.14.6",
     "Replicate": "https://github.com/nicholascelestin/replicate-js.git",
     "bcrypt": "5.1.0",
-    "cssnano": "5.1.13",
-    "framer-motion": "7.6.1",
+    "cssnano": "5.1.14",
+    "framer-motion": "7.6.6",
     "jsonwebtoken": "8.5.1",
     "lodash": "4.17.21",
-    "next": "12.3.1",
+    "next": "12.3.3",
+    "next-compose-plugins": "2.2.1",
+    "next-transpile-modules": "10.0.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hot-toast": "2.4.0",
@@ -44,8 +46,8 @@
     "@typescript-eslint/eslint-plugin": "latest",
     "@typescript-eslint/parser": "latest",
     "@vitejs/plugin-react": "2.1.0",
-    "autoprefixer": "10.4.12",
-    "eslint": "8.25.0",
+    "autoprefixer": "10.4.13",
+    "eslint": "8.27.0 ",
     "eslint-config-next": "latest",
     "eslint-config-prettier": "latest",
     "eslint-plugin-import": "latest",
@@ -56,11 +58,11 @@
     "jsdom": "20.0.1",
     "postcss": "8.4.18",
     "prettier": "2.7.1",
-    "prisma": "4.6.0",
-    "tailwindcss": "3.2.1",
+    "prisma": "4.6.1",
+    "tailwindcss": "3.2.4",
     "typescript": "4.8.4",
-    "vite": "3.1.8",
-    "vitest": "0.24.3"
+    "vite": "3.2.3",
+    "vitest": "0.25.1"
   },
   "packageManager": "yarn@3.2.4"
 }
