+++ web/src/utils/storage.ts
@@ -1,0 +1,24 @@
+import { AuthSession } from "@daydream/common/types";
+
+export function storeAuthSession(session: AuthSession) {
+  localStorage.setItem("jwt", session.jwt);
+  localStorage.setItem("userId", session.userId);
+  localStorage.setItem("userName", session.userName);
+  localStorage.setItem("userAvatar", session.userAvatar ?? "");
+}
+
+export function getAuthSession(): AuthSession {
+  return {
+    jwt: localStorage.getItem("jwt") ?? "",
+    userId: localStorage.getItem("userId") ?? "",
+    userName: localStorage.getItem("userName") ?? "",
+    userAvatar: localStorage.getItem("userAvatar"),
+  };
+}
+
+export async function clearAuthSession() {
+  localStorage.removeItem("jwt");
+  localStorage.removeItem("userId");
+  localStorage.removeItem("userName");
+  localStorage.removeItem("userAvatar");
+}
